Merge branch 'XX_Master'

Summary of the major changes:
1. Modify the jgefs_keep_data.py and jgefs_archive.py that when the folder does not exist, they will not exit with error
2. Modify jgefs_cleanup.py to clean the pgrb2ap25 instead of pgrb2a1p0
3. Modify default value of DIRS_TO_KEEP and DIRS_TO_ARCHIVE in default user configure files since pgrb2a1p0 has removed
4. Modify py/GEFS_XML_For_Tasks.py to make the python scripts can modify the depency of jgefs_post_low and jgefs_prdgen_low automatically
5. Modify to consistent with the modification of the master Revision 54e87fbb
  1) Replace FHOUTHR with FHOUTHF
  2) Add FHOUTHF and FHOUTLF to the default user configure files
  3) Add FHOUTHF and FHOUTLF to user_full.conf to be convenient for users to modify the output frequency for high/low resolution
  4) Modify py/GEFS_XML_For_Tasks.py to make the python scripts can modify the depency of jgefs_post_low and jgefs_prdgen_low automatically
6. Change the default shell using bash to force the script using bash
7. Make changes to run updated FV3 based GEFS on Theia
  1) Add jgefs_init_recenter.sh for Theia
  2) Modify setbase and gefs_config files
    a) Make consistent with the configuration for theia and cray
    b) Make the libs to point to the common libs on theia
  3) Modify user configure files for Theia
  4) Fix a small bug for task recenter on CRAY, Theia and WCOSS
  5) Added user_full_ForTest.conf for the test on Theia using Bing Fu's initial data
  6) Try to create a new bash script to unify compilling, installing and clear GEFS, as well as running rocoto
8. Modify the job scripts to generate the PDY file into the corresponding folders.
9. Fix a small bug in the following files
  1)modified: jobs/JGEFS_ENSSTAT
  2)modified: jobs/JGEFS_PRDGEN

+++ rocoto_new/bin/cray/jgefs_keep_data.py
@@ -69,17 +69,18 @@ output_path = work_dir + "/com/gens/dev/gefs." + date_string + "/" + cycle
 # Output directories
 for directory in dirs_to_keep:
 	output_dir = output_path + "/" + directory
-	destination_dir = destination_path + "/" + directory
-	if ( os.path.exists(destination_dir) ):
-		if(clobber):
-			shutil.rmtree(destination_dir)
-		else:
-			print("FATAL: Destination diretory " + destination_dir + " already exists and clobber is False")
-			quit(-102)
+	if os.path.exists(output_dir):
+		destination_dir = destination_path + "/" + directory
+		if ( os.path.exists(destination_dir) ):
+			if(clobber):
+				shutil.rmtree(destination_dir)
+			else:
+				print("FATAL: Destination diretory " + destination_dir + " already exists and clobber is False")
+				quit(-102)
 
-	if ( not os.path.exists(destination_path) ):
-		os.makedirs(destination_path)
+		if ( not os.path.exists(destination_path) ):
+			os.makedirs(destination_path)
 
-	print("Copying files for " + directory)
-	print("    From " + output_dir + " to " + destination_dir)
-	shutil.copytree(output_dir, destination_dir)
+		print("Copying files for " + directory)
+		print("    From " + output_dir + " to " + destination_dir)
+		shutil.copytree(output_dir, destination_dir)
